rename: 도메인 기반으로 pages/ 파일 그룹화 - related to #106
