+++ apps/frontend/src/pages/auth/AuthCheck.tsx
