+++ apps/frontend/src/pages/common/GlobalError.tsx
