+++ apps/frontend/src/pages/common/ServicePreparation.tsx
