+++ apps/frontend/src/pages/learn/InitialFields.tsx
