+++ apps/frontend/src/pages/quiz/QuizResult.tsx
