+++ apps/frontend/src/pages/quiz/QuizResultError.tsx
